pyproject.toml
[tool.poetry]
name = "ml_service"
version = "0.1.0"
description = "Machine Learning Service"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.30.0"
torch = "^2.0.0"
accelerate = "^0.20.0"
bitsandbytes = "^0.41.0"  # Pour la quantification 4-bit
peft = "^0.4.0"  # Pour le fine-tuning efficace
optimum = "^1.13.2"  # Optimisations pour les modèles
auto-gptq = "^0.4.2"  # Quantification automatique
httpx = "^0.24.0"
fastapi = "^0.100.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.0"
redis = "^4.5.0"
psutil = "^5.9.0"  # Pour le monitoring système
gputil = "^1.4.0"  # Pour le monitoring GPU
py3nvml = "^0.2.7"  # Pour NVIDIA GPU monitoring

# APIs et Services Gratuits
# 1. HuggingFace Inference API - Tier gratuit
# 2. OpenAI API - Tier gratuit avec nouveaux comptes
# 3. Replicate API - Crédits gratuits
# 4. Ollama - Modèles locaux gratuits
# 5. CohereAI - Tier gratuit généreux

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 